Hide "Processing Issues" setting if it was never activated
As one step towards deprecating "Legacy Reprocessing", this hides the "Processing Issues" settings page completely if it has never been active before.

It is still showing up for projects which have it active, or have ever had it active.

Along with that, it will now display a deprecation notice, though I might remove it from this PR if it is too early for that.